* wip/1.2: (21 commits)
  Added fix to ensure correct normalization and return (#588)
  Outputs time respecting backend preferences (#572)
  Bump minimum Laravel version to 9.1
  Fix site relative partial paths failing in 1.2 (#587)
  Switch back to using Laravel CacheServiceProvider
  Rebuild Snowboard agian
  Revert "Rebuild Snowboard"
  Rebuild Snowboard
  Allow a string selector for the form in a request
  Allow a string selector for the form in a request
  Use correct line breaks for Windows tests
  Backport ViewMaker tests from 1.2 branch
  Add additional testMakePartial cases to ViewMaker unit tests (#586)
  Added replaced plugins to the normalize map to ensure classloader namespace aliasing detects replacements (#585)
  Split tests into relevant module folders
  [FIX] Fix plugin flags not loading from cache correctly (#582)
  farsi spelling correction (#579)
  Moved plugin replacement namespace aliasing into register replacement method (#580)
  fix typo
  Add Winter 1.2 as version option in bug report
  ...

+++ modules/system/helpers/Cache.php
@@ -36,7 +36,11 @@ class Cache
      */
     public function clearCombiner()
     {
-        foreach (File::directories(storage_path().'/cms/combiner') as $directory) {
+        $path = storage_path().'/cms/combiner';
+        if (!is_dir($path)) {
+            return;
+        }
+        foreach (File::directories($path) as $directory) {
             File::deleteDirectory($directory);
         }
     }
@@ -46,6 +50,10 @@ class Cache
      */
     public function clearCache()
     {
+        $path = storage_path().'/cms/cache';
+        if (!is_dir($path)) {
+            return;
+        }
         foreach (File::directories(storage_path().'/cms/cache') as $directory) {
             File::deleteDirectory($directory);
         }
@@ -56,6 +64,10 @@ class Cache
      */
     public function clearTwig()
     {
+        $path = storage_path().'/cms/twig';
+        if (!is_dir($path)) {
+            return;
+        }
         foreach (File::directories(storage_path().'/cms/twig') as $directory) {
             File::deleteDirectory($directory);
         }
@@ -66,7 +78,6 @@ class Cache
      */
     public function clearMeta()
     {
-        File::delete(storage_path().'/cms/disabled.json');
         File::delete(App::getCachedCompilePath());
         File::delete(App::getCachedServicesPath());
     }
